Fix de varias funciones.

+++ grage-pfs/app/includes/pfs-fat32.h
@@ -105,7 +105,7 @@
 	int8_t pfs_fat32_unlink_FatEntryChain(Volume * v , FatFile * fd);
 	void pfs_fat32_unlink(Volume * v , FatFile * fd);
 	void pfs_fat32_rmdir(Volume * v , FatFile * fd);
-	uint16_t pfs_fat32_read(Volume * , FatFile * , char * , size_t);
+	uint32_t pfs_fat32_read(Volume * , FatFile * , char * , size_t);
 
 //UTILS
 	Volume * pfs_fat_utils_loadVolume( BPB * b );
